# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VideoForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/export:
    get:
      operationId: exportProject
      tags: [projects]
      summary: Export project as JSON
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExport"
...
  /projects/{projectId}/render:
    get:
      operationId: getRenderStatus
      tags: [render]
      summary: Get render status for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Render status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderJob"
    post:
      operationId: startRender
      tags: [render]
      summary: Start a render job
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderInput"
      responses:
        "200":
          description: Render job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderJob"
  /projects/{projectId}/render/retry:
    post:
      operationId: retryRender
      tags: [render]
      summary: Retry a failed render
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Render job restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderJob"
  /projects/{projectId}/publish:
    get:
      operationId: getPublishInfo
      tags: [publish]
      summary: Get publish status for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Publish info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishInfo"
    post:
      operationId: publishToYoutube
      tags: [publish]
      summary: Upload to YouTube or prepare download
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishInput"
      responses:
        "200":
          description: Publish result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, analyzing, concepts, scripting, media, voiceover, rendering, ready, published]
        referenceUrl:
          type: ["string", "null"]
        referenceTopic:
          type: ["string", "null"]
        selectedConceptId:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        renderProgress:
          type: ["integer", "null"]
        renderStatus:
          type: ["string", "null"]
        outputFormats:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        referenceUrl:
          type: string
        referenceTopic:
          type: string
      required:
        - title
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        thumbnailUrl:
          type: string
    ProjectExport:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
        script:
          $ref: "#/components/schemas/Script"
        concepts:
          type: array
          items:
            $ref: "#/components/schemas/Concept"
        captions:
          $ref: "#/components/schemas/Captions"
        mediaAssets:
          type: array
          items:
            $ref: "#/components/schemas/MediaAsset"
        rightsManifest:
          type: object
          additionalProperties: true
        attributionText:
          type: string
        exportedAt:
          type: string
      required:
        - project
        - exportedAt
...
    Concept:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        angle:
          type: string
        hook:
          type: string
        structure:
          type: string
        targetAudience:
          type: string
        estimatedDuration:
          type: integer
        uniqueElements:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - angle
        - hook
        - structure
        - targetAudience
        - estimatedDuration
        - uniqueElements
        - createdAt
...
    Script:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        content:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ScriptSection"
        wordCount:
          type: integer
        estimatedDuration:
          type: integer
        updatedAt:
          type: string
        originalityScore:
          type: ["integer", "null"]
        isMock:
          type: boolean
      required:
        - id
        - projectId
        - content
        - sections
        - wordCount
        - estimatedDuration
        - updatedAt
    ScriptSection:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [hook, intro, main, cta, outro]
        title:
          type: string
        content:
          type: string
        durationSeconds:
          type: integer
      required:
        - id
        - type
        - title
        - content
        - durationSeconds
...
    MediaAsset:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          enum: [video, image, audio, stock_video, stock_photo]
        source:
          type: string
          enum: [uploaded, pexels, generated]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        filename:
          type: string
        duration:
          type: ["integer", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        attribution:
          type: ["string", "null"]
        license:
          type: ["string", "null"]
        createdAt:
          type: string
        orderIndex:
          type: integer
      required:
        - id
        - projectId
        - type
        - source
        - url
        - filename
        - createdAt
        - orderIndex
...
    Captions:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        srtContent:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CaptionEntry"
        language:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - srtContent
        - entries
        - language
        - createdAt
    CaptionEntry:
      type: object
      properties:
        index:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        text:
          type: string
      required:
        - index
        - startTime
        - endTime
        - text
    RenderJob:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          type: string
          enum: [queued, rendering, done, error]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        format:
          type: string
          enum: [landscape, portrait, both]
        landscapeUrl:
          type: ["string", "null"]
        portraitUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        isMock:
          type: boolean
      required:
        - id
        - projectId
        - status
        - progress
        - format
        - createdAt
    RenderInput:
      type: object
      properties:
        format:
          type: string
          enum: [landscape, portrait, both]
          default: both
        includeSubtitles:
          type: boolean
          default: true
      required:
        - format
    PublishInfo:
      type: object
      properties:
        projectId:
          type: string
        youtubeEnabled:
          type: boolean
        lastPublishedAt:
          type: ["string", "null"]
        youtubeVideoId:
          type: ["string", "null"]
        downloadUrls:
          type: object
          properties:
            landscape:
              type: ["string", "null"]
            portrait:
              type: ["string", "null"]
            srt:
              type: ["string", "null"]
            thumbnail:
              type: ["string", "null"]
            projectJson:
              type: ["string", "null"]
            attribution:
              type: ["string", "null"]
          required: []
      required:
        - projectId
        - youtubeEnabled
        - downloadUrls
    PublishInput:
      type: object
      properties:
        platform:
          type: string
          enum: [youtube, download]
          default: download
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - platform
    PublishResult:
      type: object
      properties:
        success:
          type: boolean
        platform:
          type: string
        youtubeVideoId:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        downloadUrls:
          type: object
          additionalProperties: true
        message:
          type: string
      required:
        - success
        - platform
        - message
...