# lib/api-zod/src/generated/types/scriptSection.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * VideoForge API
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptSectionType } from './scriptSectionType';

export interface ScriptSection {
  id: string;
  type: ScriptSectionType;
  title: string;
  content: string;
  durationSeconds: number;
}